Blog Tags
Added tags to blogs. Because of the supabase structure, each blog can only have one tag. Ideally this will get fixed by Vincent sometime.

If we fix the supabase structure, that actually might result in some errors. For example in "blog/edit.html" and "blog/new.html" in lines ~30. These are the lines that reference post.category.name. It's worth noting this may be affected by changes to supabase